Rework babySplitListener into MonoplexingListener (#25)
This adds IngressListener which allows a listener to be given to a direct
instantiation of MonoplexingListener to have its connections come out
the other side. By adding multiple such listeners, it effectively acts
as a monoplexer. This doesn't really change how the babySplitListener
worked, it just leans into the fact that it serializes down a channel
with a Listener interface on the other end.

Because of this change we can no longer close the channel (which is
fine); this means simplifying some of the other control logic.