Revamp server-led mechanism
This replaces the previous mechanism of generating the full credential
set and returning it with creating and storing an activation token that
can be passed in as the nonce in a fetch request.

This allows the mechanism on the node side to be essentially the same,
including all of the private keys being generated node-side. Instead of
generating and signing a nonce on its own, it signs the activation token
as the nonce.