Add "wrapping" registration flow
This uses a KMS (a wrapper) as a third-party arbitrator of validity. The
signed bundle can contain registration information encrypted by a shared
KMS. This information consists of the nonce and the public key from the
signer of the bundle itself, which means our existing protections
against replay (such as the result being encrypted to the signed
encryption key ID) apply here as well. If the value is able to be
successfully decrypted, the node is authorized.

For multi-hop situations, an interim node can unwrap the value using its
wrapper and encrypt the message to the server using its encryption key.
Notably, this encrypted value is placed outside the signed bundle. If
the server trusts this node it can verify that the node provided the
unwrapped values by decrypting using its shared key. Notably, it is
still on the server to validate the contained information (nonce and
key) after it validates the signature of the bundle.