Don't allow certificate lifetime to be set zero (#44)
This is almost certainly not what a caller intended to do and instead is
likely because the caller meant to simply use the default lifetime. So,
specify zero as the default lifetime; if you truly want it short,
setting it to literally anything other than zero will do what was
desired.