Allow multiplexing listeners to close
Currently multiplexing listeners sit forever watching the incoming
channel to ensure it's drained once the context is canceled. In normal
usage this is fine, however it creates oodles of never-exiting
goroutines in tests that make it hard to follow traces and chew up
memory and context switches.

The reason it was left this way was to be safe to ensure we don't send
on a closed channel or run the risk of closing an already closed
channel. The changes here should be safe on both counts; closing the
channel is protected by a sync.Once, and any send on the channel is in a
mutex-protected critical section that checks whether the channel has
been closed.

Tests pass (and they did catch an earlier panic before I updated
GetListener to not also close the channel) and with some log output
added, testing manually within Boundary also passes, and I can verify
the function now exits.